Fixed json merging for synch command
Before key can be updated (replaced) in json structure assuming
synchronization is required.

It was recognized through existence of meta-data file, which
meant that skeleton was initialized and is either updated with
preceding checks or validated. Merged json file that needs to
be synchronized in initialized package would be generated with
assumend correct field order.

Current solution is based on command name, so argument parsing
needs to precede MergedJsonFactory instantiation.